refactor : 그룹의 모든 회원 조회 API 멤버에서 그룹 도메인으로 이동 (#75)
