+++ src/main/java/com/soma/snackexercise/controller/exgroup/ExgroupController.java
@@ -45,6 +45,13 @@ public class ExgroupController {
         return Response.success(exGroupService.findGroup(groupId));
     }
 
+    @Operation(summary = "하나의 운동 그룹에 속한 모든 회원 조회", description = "하나의 운동 그룹에 속한 모든 회원을 조회합니다.", security = { @SecurityRequirement(name = "bearer-key") })
+    @Parameter(name = "groupId", description = "조회할 운동 그룹 ID")
+    @GetMapping("/exgroups/{groupId}/members")
+    public Response getAllExgroupMembers(@PathVariable("groupId") Long groupId){
+        return Response.success(exGroupService.getAllExgroupMembers(groupId));
+    }
+
     @Operation(summary = "운동 그룹 수정",
             description = "하나의 운동 그룹을 수정합니다.",
             security = { @SecurityRequirement(name = "bearer-key") },
