+++ src/main/java/com/soma/snackexercise/controller/member/MemberController.java
@@ -28,14 +28,6 @@ public class MemberController {
         return "jwtTest 요청 성공";
     }
 
-
-    @Operation(summary = "하나의 운동 그룹에 속한 모든 회원 조회", description = "하나의 운동 그룹에 속한 모든 회원을 조회합니다.", security = { @SecurityRequirement(name = "bearer-key") })
-    @Parameter(name = "groupId", description = "조회할 운동 그룹 ID")
-    @GetMapping("/exgroups/{groupId}/members")
-    public Response getAllExgroupMembers(@PathVariable("groupId") Long groupId){
-        return Response.success(memberService.getAllExgroupMembers(groupId));
-    }
-
     @Operation(summary = "멤버 수정",
             description = "멤버 한명을 수정합니다.",
             security = { @SecurityRequirement(name = "bearer-key") },
